Refactor define and uglify configuration plugins
